Line-back-end/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
aenum==3.1.16
aiohttp==3.9.5
aiosignal==1.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
fastapi==0.119.0
frozenlist==1.8.0
future==1.0.0
groq==0.32.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
line-bot-sdk==3.11.0
multidict==6.7.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
propcache==0.4.1
requests==2.31.0
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
langchain==1.0.0
langchain-core==1.0.0
langchain-groq==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37